requirements.txt
pydantic>=2.6,<3
httpx>=0.27,<1
python-dotenv>=1.0,<2
rank-bm25>=0.2.2,<1
rich>=13.7,<14
